pyproject.toml
[project]
name = "book-voice-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "gtts>=2.5.4",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
]
